# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Refactoring",
        "title": "OnlineJudge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Eutop1a",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:65533",
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "user_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/controller._LoginSuccess"
                        }
                    },
                    "400": {
                        "description": "用户名不存在或验证码错误",
                        "schema": {
                            "$ref": "#/definitions/controller._LoginError"
                        }
                    },
                    "401": {
                        "description": "验证码过期",
                        "schema": {
                            "$ref": "#/definitions/controller._LoginError"
                        }
                    },
                    "403": {
                        "description": "密码错误",
                        "schema": {
                            "$ref": "#/definitions/controller._LoginError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._LoginError"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "user_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/controller._RegisterSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller._RegisterError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/controller._RegisterError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller._RegisterError"
                        }
                    }
                }
            }
        },
        "/send-email-code": {
            "post": {
                "description": "发送验证码接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发送验证码成功",
                        "schema": {
                            "$ref": "#/definitions/controller._SendCodeSuccess"
                        }
                    },
                    "400": {
                        "description": "邮箱格式错误",
                        "schema": {
                            "$ref": "#/definitions/controller._SendCodeError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._SendCodeError"
                        }
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "description": "获取用户详细信息接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取用户详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/controller._GetUserDetailSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._GetUserDetailError"
                        }
                    },
                    "403": {
                        "description": "没有此用户ID",
                        "schema": {
                            "$ref": "#/definitions/controller._GetUserDetailError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._GetUserDetailError"
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户详细信息接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新用户详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "用户邮箱",
                        "name": "email",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/controller._UpdateUserDetailSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._UpdateUserDetailError"
                        }
                    },
                    "403": {
                        "description": "没有此用户ID or 验证码错误",
                        "schema": {
                            "$ref": "#/definitions/controller._UpdateUserDetailError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._UpdateUserDetailError"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除用户接口",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/controller._DeleteUserSuccess"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/controller._DeleteUserError"
                        }
                    },
                    "403": {
                        "description": "没有此用户ID",
                        "schema": {
                            "$ref": "#/definitions/controller._DeleteUserError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/controller._DeleteUserError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller._DeleteUserError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": "delete user error"
                }
            }
        },
        "controller._DeleteUserSuccess": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "format": "string",
                    "example": "success delete user"
                }
            }
        },
        "controller._GetUserDetailError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": "get user detail error"
                }
            }
        },
        "controller._GetUserDetailSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/mysql.User"
                },
                "msg": {
                    "type": "string",
                    "format": "string",
                    "example": "success get user detail"
                }
            }
        },
        "controller._LoginError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": "invalidate email format"
                }
            }
        },
        "controller._LoginSuccess": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "format": "string",
                    "example": "login successfully"
                }
            }
        },
        "controller._RegisterError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": "register error"
                }
            }
        },
        "controller._RegisterSuccess": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "format": "string",
                    "example": "registration successful"
                },
                "token": {
                    "type": "string",
                    "format": "token string",
                    "example": "bearer token"
                }
            }
        },
...
        "controller._SendCodeError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": "invalidate email format"
                }
            }
        },
        "controller._SendCodeSuccess": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "format": "string",
                    "example": "send verification code successfully"
                }
            }
        },
        "controller._UpdateUserDetailError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string",
                    "example": "update user information error"
                }
            }
        },
        "controller._UpdateUserDetailSuccess": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "format": "string",
                    "example": "success update user information"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "mysql.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "last_login_data": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "registration_date": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    }
}